+++ packages/atproto/lib/atproto.dart
@@ -34,6 +34,7 @@ export 'package:atproto/src/entities/repo_blocks.dart';
 export 'package:atproto/src/entities/repo_commit.dart';
 export 'package:atproto/src/entities/repo_commit_paths.dart';
 export 'package:atproto/src/entities/repo_commits.dart';
+export 'package:atproto/src/entities/repo_head.dart';
 export 'package:atproto/src/entities/repo_op.dart';
 export 'package:atproto/src/entities/repo_ref.dart';
 export 'package:atproto/src/entities/report.dart';